Evolution in time is set up, we miss the interface to allocate and release the resources. The code does not compile, but it is always better that Matteo's hairstyle.

+++ resources.h
@@ -45,13 +45,14 @@ public:
      */
     void res_allocate(Wlk_Resources const& needed, Wlk_Resources & occup) {
         /* Dimension checks */
-        assert(needed.ntypes() == occup.ntypes() && needed.ntypes() == available.size());
+      //      if (needed.ntypes() == occup.ntypes() && needed.ntypes() == (int)available.size()) 
+            assert(needed.ntypes() == occup.ntypes() && needed.ntypes() == (int)available.size());
         //if (needed.size() != occup.size() ||
         //        needed.size() != available.size() ) {
 
         //    cout << "Input vectors to function 'allocate' in Resource class have wrong length" << endl;
         //}
-
+      //      cout<< (int)available.size() << " " << needed.ntypes()  << " " <<  occup.ntypes()<< endl;
         cout << "Allocated the following resources:" << endl;
 
         /* Go through all the types of resources */
